split files again
